+++ InfoService/trunk/src/jap/MixCascadeUpdater.java
@@ -45,7 +45,7 @@ import logging.LogType;
  */
 public class MixCascadeUpdater extends AbstractDatabaseUpdater
 {
-	private static final int UPDATE_INTERVAL_MS = 5 * 60000;
+	private static final long UPDATE_INTERVAL_MS = 5 * 60000l;
 
 	public MixCascadeUpdater()
 	{
